+++ packages/orq-rc/RELEASES.md
@@ -10559,3 +10559,13 @@ Based on:
 - [python v4.15.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.7 - packages/orq-rc
+
+## 2026-08-31 11:27:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.2 (2.933.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.8 - packages/orq-rc